requirements.txt
flask==2.3.3
rembg==2.0.50
pillow==10.0.1
gunicorn==21.2.0

